# server-player/internal/pb/swagger/players_api/players.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "players_api/players.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PlayersService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/player/quiz/finish": {
      "put": {
        "operationId": "PlayersService_SetResultsByQuiz",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1SetResultsByQuizResponce"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "tgId",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "themaId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "score",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "time",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "PlayersService"
        ]
      }
    },
    "/api/player/tenquestions/{topicName}": {
      "get": {
        "operationId": "PlayersService_GetTenQuestionsByTopic",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetTenQuestionsByTopicResponce"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "topicName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "PlayersService"
        ]
      }
    },
    "/api/users/topics": {
      "get": {
        "operationId": "PlayersService_GetAllTopics",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAllTopicsResponce"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "PlayersService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1GetAllTopicsResponce": {
      "type": "object",
      "properties": {
        "topics": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Topic"
          }
        }
      }
    },
    "v1GetTenQuestionsByTopicResponce": {
      "type": "object",
      "properties": {
        "questions": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Question"
          }
        },
        "topicId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1Question": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "correctIndex": {
          "type": "string",
          "format": "int64"
        },
        "level": {
          "type": "string"
        }
      }
    },
    "v1SetResultsByQuizResponce": {
      "type": "object"
    },
    "v1Topic": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "title": {
          "type": "string"
        }
      }
    }
  }
}